moved some generic components into lib folder
